Change how analyses are turned on/off ('generate')
With this merge, there is a single list 'generate' in the
'[output]' section that determines which analyses are run.  This
option replaces the 'generate' options in in the sections for
each specific analysis module.

The syntax for generate is described in config.template as follows.

'generate' is a list of which analyses to generate.  Valid names are:
   'ohc_timeseries', 'sst_timeseries', 'sst_modelvsobs', 'sss_modelvsobs',
   'mld_modelvsobs', 'seaice_timeseries', 'seaice_modelvsobs'
 the following shortcuts exist:
   'all' -- all analyses will be run
   'all_timeseries' -- all time-series analyses will be run
   'all_modelvsobs' -- all analyses comparing model climatologies with
                       observatoins will be run
   'all_ocean' -- all ocean analyses will be run
   'all_seaice' -- all sea-ice analyses will be run
   'no_ohc_timeseries' -- skip the 'ohc_timeseries' (and similarly with the other analyses).
   'no_ocean', 'no_timeseries', etc. -- in analogy to 'all_*', skip the given category of analysis

With a command line argument `--generate`, a comma-separated list
can be used to select which analyses are performed.  The command-line
argument overrides the config option.